Rename lib files as ('lowercase_with_underscores')
